refactor(peeroxide-dht): visibility reform Phase 1 (promote 7 modules, demote 5)
Per VISIBILITY_POLICY.md u00a711.2 + u00a711.3 (verified Phase 0C):

PROMOTE (drop #[doc(hidden)], add module doc stubs): holepuncher, io, peer,

persistent, secretstream, secure_payload, socket_pool.

DEMOTE (pub(crate)): compact_encoding, nat, query, router, routing_table.

Cascade fixes:

- Holepuncher.nat field demoted to pub(crate)

- blind_relay encode_*/decode_* free fns demoted to pub(crate)

- pub use re-exports added at crate root for EncodingError, QueryReply,

  QueryResult, Router, HandshakeAction, HolepunchAction, ForwardEntry,

  HandshakeResult, HolepunchResult, RouterError to preserve public method signatures

Test relocations:

- tests/golden_interop.rs and tests/blind_relay_golden_interop.rs moved to

  #[cfg(test)] modules co-located with implementations (required by module demotion).

Gates: cargo build, cargo test --workspace, cargo clippy -- -D warnings all clean.
