+++ peeroxide-dht/src/router.rs
@@ -1,3 +1,5 @@
+#![allow(missing_docs)]
+
 use std::collections::HashMap;
 use std::time::{Duration, Instant};
