--- peeroxide-dht/tests/golden_interop.rs
@@ -1,475 +1,0 @@
-//! Golden fixture interop tests — verifies our compact-encoding Rust implementation
-//! produces byte-identical output to the Node.js `compact-encoding` npm package.
-//!
-//! The golden fixtures are generated by `tests/node/generate-golden.js`.
-//! Run `node generate-golden.js` in `tests/node/` to regenerate.
-
-use peeroxide_dht::compact_encoding::*;
-use serde::Deserialize;
-
-#[derive(Deserialize)]
-struct GoldenFile {
-    #[allow(dead_code)]
-    generated_by: String,
-    #[allow(dead_code)]
-    version: String,
-    fixtures: Vec<Fixture>,
-}
-
-#[derive(Deserialize)]
-struct Fixture {
-    #[serde(rename = "type")]
-    typ: String,
-    label: String,
-    value: serde_json::Value,
-    hex: String,
-}
-
-fn load_fixtures() -> Vec<Fixture> {
-    let path = concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../tests/interop/golden-fixtures.json"
-    );
-    let data = std::fs::read_to_string(path)
-        .unwrap_or_else(|e| panic!("Failed to read golden fixtures at {path}: {e}. Run `node generate-golden.js` in tests/node/ first."));
-    let file: GoldenFile = serde_json::from_str(&data)
-        .unwrap_or_else(|e| panic!("Failed to parse golden fixtures: {e}"));
-    file.fixtures
-}
-
-fn expected_bytes(hex_str: &str) -> Vec<u8> {
-    hex::decode(hex_str).unwrap_or_else(|e| panic!("Invalid hex '{hex_str}': {e}"))
-}
-
-fn encode_with_pre(pre_fn: impl FnOnce(&mut State), enc_fn: impl FnOnce(&mut State)) -> Vec<u8> {
-    let mut state = State::new();
-    pre_fn(&mut state);
-    state.alloc();
-    enc_fn(&mut state);
-    state.buffer
-}
-
-fn val_u64(v: &serde_json::Value) -> u64 {
-    match v {
-        serde_json::Value::Number(n) => n.as_u64().unwrap(),
-        serde_json::Value::String(s) => s.parse::<u64>().unwrap(),
-        _ => panic!("Expected number or string for u64, got {v:?}"),
-    }
-}
-
-fn val_i64(v: &serde_json::Value) -> i64 {
-    match v {
-        serde_json::Value::Number(n) => n.as_i64().unwrap(),
-        serde_json::Value::String(s) => s.parse::<i64>().unwrap(),
-        _ => panic!("Expected number or string for i64, got {v:?}"),
-    }
-}
-
-fn val_f64(v: &serde_json::Value) -> f64 {
-    v.as_f64().unwrap()
-}
-
-fn val_f32(v: &serde_json::Value) -> f32 {
-    v.as_f64().unwrap() as f32
-}
-
-fn val_str(v: &serde_json::Value) -> &str {
-    v.as_str().unwrap()
-}
-
-fn val_bool(v: &serde_json::Value) -> bool {
-    v.as_bool().unwrap()
-}
-
-#[test]
-fn golden_uint8() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint8") {
-        let val = val_u64(&f.value) as u8;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint8(s, val), |s| encode_uint8(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint8(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}: expected {val}, got {decoded}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint16() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint16") {
-        let val = val_u64(&f.value) as u16;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint16(s, val), |s| encode_uint16(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint16(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint24() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint24") {
-        let val = val_u64(&f.value) as u32;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint24(s, val), |s| encode_uint24(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint24(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint32() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint32") {
-        let val = val_u64(&f.value) as u32;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint32(s, val), |s| encode_uint32(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint32(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint64() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint64") {
-        let val = val_u64(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint64(s, val), |s| encode_uint64(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint64(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint_varint() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint") {
-        let val = val_u64(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_uint(s, val), |s| encode_uint(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_int_zigzag() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "int") {
-        let val = val_i64(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_int(s, val), |s| encode_int(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_int(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_int8() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "int8") {
-        let val = val_i64(&f.value) as i8;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_int8(s, val), |s| encode_int8(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_int8(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_int16() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "int16") {
-        let val = val_i64(&f.value) as i16;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_int16(s, val), |s| encode_int16(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_int16(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_int32() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "int32") {
-        let val = val_i64(&f.value) as i32;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_int32(s, val), |s| encode_int32(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_int32(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_int64() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "int64") {
-        let val = val_i64(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_int64(s, val), |s| encode_int64(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_int64(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_float32() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "float32") {
-        let val = val_f32(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_float32(s, val), |s| encode_float32(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_float32(&mut state).unwrap();
-        assert_eq!(decoded.to_bits(), val.to_bits(), "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_float64() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "float64") {
-        let val = val_f64(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_float64(s, val), |s| encode_float64(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_float64(&mut state).unwrap();
-        assert_eq!(decoded.to_bits(), val.to_bits(), "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_bool() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "bool") {
-        let val = val_bool(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_bool(s, val), |s| encode_bool(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_bool(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_buffer() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "buffer") {
-        let expected = expected_bytes(&f.hex);
-
-        let val: Option<Vec<u8>> = if f.value.is_null() {
-            None
-        } else {
-            let hex_str = val_str(&f.value);
-            if hex_str.is_empty() {
-                // Node.js encodes both null and empty buffer as 0x00
-            None
-            } else {
-                Some(hex::decode(hex_str).unwrap())
-            }
-        };
-
-        let val_slice: Option<&[u8]> = val.as_deref();
-        let encoded = encode_with_pre(
-            |s| preencode_buffer(s, val_slice),
-            |s| encode_buffer(s, val_slice),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_buffer(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_string() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "string") {
-        let val = val_str(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(|s| preencode_string(s, val), |s| encode_string(s, val));
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_string(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_fixed32() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "fixed32") {
-        let val_vec = hex::decode(val_str(&f.value)).unwrap();
-        let val: [u8; 32] = val_vec.try_into().unwrap();
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| { preencode_fixed32(s, &val).unwrap(); },
-            |s| encode_fixed32(s, &val),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_fixed32(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_fixed64() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "fixed64") {
-        let val_vec = hex::decode(val_str(&f.value)).unwrap();
-        let val: [u8; 64] = val_vec.try_into().unwrap();
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| { preencode_fixed64(s, &val).unwrap(); },
-            |s| encode_fixed64(s, &val),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_fixed64(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_ipv4() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "ipv4") {
-        let val = val_str(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_ipv4(s, val),
-            |s| encode_ipv4(s, val).unwrap(),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_ipv4(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_ipv6() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "ipv6") {
-        let val = val_str(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_ipv6(s, val),
-            |s| encode_ipv6(s, val).unwrap(),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        // Our decoder outputs expanded form; compare via Ipv6Addr for canonical equivalence
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_ipv6(&mut state).unwrap();
-        let expected_addr: std::net::Ipv6Addr = val.parse().unwrap();
-        let decoded_addr: std::net::Ipv6Addr = decoded.parse().unwrap();
-        assert_eq!(decoded_addr, expected_addr, "DECODE {}: expected {expected_addr}, got {decoded_addr}", f.label);
-    }
-}
-
-#[test]
-fn golden_ip_dual() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "ip") {
-        let val = val_str(&f.value);
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_ip(s, val),
-            |s| encode_ip(s, val).unwrap(),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_ip(&mut state).unwrap();
-
-        // IPv6 needs canonical comparison since our decoder fully expands addresses
-        if val.contains(':') {
-            let expected_addr: std::net::Ipv6Addr = val.parse().unwrap();
-            let decoded_addr: std::net::Ipv6Addr = decoded.parse().unwrap();
-            assert_eq!(decoded_addr, expected_addr, "DECODE {}", f.label);
-        } else {
-            assert_eq!(decoded, val, "DECODE {}", f.label);
-        }
-    }
-}
-
-#[test]
-fn golden_ipv4_address() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "ipv4Address") {
-        let obj = f.value.as_object().unwrap();
-        let host = obj["host"].as_str().unwrap();
-        let port = obj["port"].as_u64().unwrap() as u16;
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_ipv4_address(s, host, port),
-            |s| encode_ipv4_address(s, host, port).unwrap(),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let (decoded_host, decoded_port) = decode_ipv4_address(&mut state).unwrap();
-        assert_eq!(decoded_host, host, "DECODE host {}", f.label);
-        assert_eq!(decoded_port, port, "DECODE port {}", f.label);
-    }
-}
-
-#[test]
-fn golden_uint_array() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "uint_array") {
-        let val: Vec<u64> = f.value.as_array().unwrap().iter().map(val_u64).collect();
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_uint_array(s, &val),
-            |s| encode_uint_array(s, &val),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_uint_array(&mut state).unwrap();
-        assert_eq!(decoded, val, "DECODE {}", f.label);
-    }
-}
-
-#[test]
-fn golden_string_array() {
-    for f in load_fixtures().iter().filter(|f| f.typ == "string_array") {
-        let val: Vec<&str> = f.value.as_array().unwrap().iter().map(|v| v.as_str().unwrap()).collect();
-        let expected = expected_bytes(&f.hex);
-        let encoded = encode_with_pre(
-            |s| preencode_string_array(s, &val),
-            |s| encode_string_array(s, &val),
-        );
-        assert_eq!(encoded, expected, "ENCODE {}: expected {}, got {}", f.label, f.hex, hex::encode(&encoded));
-
-        let mut state = State::from_buffer(&expected);
-        let decoded = decode_string_array(&mut state).unwrap();
-        let val_owned: Vec<String> = val.iter().map(|s| s.to_string()).collect();
-        assert_eq!(decoded, val_owned, "DECODE {}", f.label);
-    }
-}
